+++ esnfed/__init__.py
@@ -65,4 +65,4 @@ __all__ = [
     "streaming",
 ]
 
-__version__ = "1.7.0"
+__version__ = "1.7.1"